jpa & service to another models -added

+++ src/main/java/com/example/meetexApi/repository/ActivitiesRepository.java
@@ -3,5 +3,5 @@ package com.example.meetexApi.repository;
 import com.example.meetexApi.model.Activities;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ActivitiesRepository extends JpaRepository<Activities, Integer> {
+public interface ActivitiesRepository extends JpaRepository<Activities, Long> {
 }
