+++ src/main/java/com/example/meetexApi/repository/PagesRepository.java
@@ -3,5 +3,5 @@ package com.example.meetexApi.repository;
 import com.example.meetexApi.model.Pages;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PagesRepository extends JpaRepository<Pages, Integer> {
+public interface PagesRepository extends JpaRepository<Pages, Long> {
 }
